Re-organize extension-lib-form.wdio-spec.js
