+++ tests/e2e/default/enketo/extension-lib-form.wdio-spec.js
@@ -1,18 +1,17 @@
-const extensionLibsPage = require('@page-objects/default/enketo/extension-lib.wdio.page');
 const commonPage = require('@page-objects/default/common/common.wdio.page');
 const utils = require('@utils');
-const userData = require('@page-objects/default/users/user.data');
 const loginPage = require('@page-objects/default/login/login.wdio.page');
 const commonEnketoPage = require('@page-objects/default/enketo/common-enketo.wdio.page');
+const genericForm = require('@page-objects/default/enketo/generic-form.wdio.page');
+const { extensionLibDoc } = require('@page-objects/default/enketo/custom-doc.wdio.page');
 
 describe('Extension lib xpath function', () => {
-  const { userContactDoc, docs } = userData;
 
   before(async () => {
-    await utils.saveDocs(docs);
+    await commonEnketoPage.uploadForm('extension-lib-average-calculator');
+    await utils.saveDoc(extensionLibDoc);
 
     const waitForServiceWorker = await utils.waitForApiLogs(utils.SW_SUCCESSFUL_REGEX);
-    await extensionLibsPage.configure(userContactDoc);
     await waitForServiceWorker.promise;
     await browser.reloadSession();
     await browser.url('/');
@@ -22,14 +21,13 @@ describe('Extension lib xpath function', () => {
     await loginPage.cookieLogin();
     await commonPage.goToReports();
 
-    await commonPage.openFastActionReport(extensionLibsPage.INTERNAL_ID, false);
+    await commonPage.openFastActionReport(/*extensionLibsPage.INTERNAL_ID*/'extension-lib-average-calculator', false);
 
     await commonEnketoPage.setInputValue('first', 5);
     await commonEnketoPage.setInputValue('second', 8);
-    await extensionLibsPage.blur();
+    await genericForm.formTitle().click();
 
-    const average = await extensionLibsPage.getAverage();
-    expect(average).to.equal('6.5');
+    expect(await commonEnketoPage.getInputValue('avg')).to.equal('6.5');
 
     // no need to submit - if the value is updated then we're good to go
   });
